flattens file structure
