+++ src/shells/interaction_wrappers.rs
